% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{catppuccin_pal}
\alias{catppuccin_pal}
\title{Islamic palette with ramped colors}
\usage{
catppuccin_pal(palette = "mocha", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from \code{\link{catppuccin_palettes}} list.}

\item{alpha}{Transparency.}

\item{reverse}{If TRUE, the direction of the colors is reversed.}
}
\description{
Islamic palette with ramped colors
}
\examples{
library(scales)
show_col(catppuccin_pal()(10))

filled.contour(volcano, color.palette = catppuccin_pal(), asp = 1)
}
