% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{catppuccin_palettes}
\alias{catppuccin_palettes}
\title{A list of color palettes from Catppuccin
Catppuccin is a community-driven pastel theme that aims to be the middle ground between low and high contrast themes. It consists of 4
soothing warm palettes with 26 eye-candy colors each, perfect for coding, designing, and much more! In addition, this repository tracks the
development of the actual color palette, the project's documentation, organization-wide assets, resources and code samples for
maintainers/developers.}
\format{
An object of class \code{list} of length 4.
}
\usage{
catppuccin_palettes
}
\description{
A list of color palettes from Catppuccin
Catppuccin is a community-driven pastel theme that aims to be the middle ground between low and high contrast themes. It consists of 4
soothing warm palettes with 26 eye-candy colors each, perfect for coding, designing, and much more! In addition, this repository tracks the
development of the actual color palette, the project's documentation, organization-wide assets, resources and code samples for
maintainers/developers.
}
\keyword{datasets}
