% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_catppuccin}
\alias{scale_color_catppuccin}
\alias{scale_colour_catppuccin}
\title{Setup color palette for ggplot2}
\usage{
scale_color_catppuccin(
  ...,
  palette = "mocha",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE
)

scale_colour_catppuccin(
  ...,
  palette = "mocha",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{...}{Additional arguments to pass to \code{\link[ggplot2]{scale_color_gradientn}}.}

\item{palette}{Choose from \code{\link{catppuccin_palettes}} list.}

\item{discrete}{Whether to use a discrete color palette.}

\item{alpha}{Transparency.}

\item{reverse}{Logical, whether to reverse the order of the colors.}
}
\description{
Setup color palette for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl))) +
  scale_colour_catppuccin(palette = "ottoman")
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = hp)) +
  scale_colour_catppuccin(palette = "shiraz", discrete = FALSE)
ggplot(data = mpg) +
  geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
  scale_colour_catppuccin(palette = "jerusalem")
ggplot(diamonds) +
  geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_catppuccin()
}
