% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_catppuccin}
\alias{scale_fill_catppuccin}
\title{Setup fill palette for ggplot2}
\usage{
scale_fill_catppuccin(
  ...,
  palette = "mocha",
  discrete = TRUE,
  alpha = 1,
  reverse = TRUE
)
}
\arguments{
\item{...}{Additional arguments to pass to \code{\link[ggplot2]{scale_fill_gradientn}}.}

\item{palette}{Choose from \code{\link{catppuccin_palettes}} list.}

\item{discrete}{Whether to use a discrete color palette.}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{reverse}{If TRUE, the direction of the colors is reversed.}
}
\description{
Setup fill palette for ggplot2
}
